Добавлен CommandManager
