+++ libs/CommandManager.cs
@@ -1,0 +1,59 @@
+﻿using Libs;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net.Sockets;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Libs
+{
+    public class CommandManager
+    {
+        public enum Commands
+        {
+            NULL,
+            SubsidiaryAdd,
+            SubsidiaryLoad,
+        }
+
+
+        public static void CommandHandler(NetworkStream stream, string input, string clientIp)
+        {
+            string message = NetManager.GetMessage(input);
+            string command = NetManager.GetCommand(input);
+
+            Console.WriteLine("{0}: {1}({2});", clientIp, command, message);
+
+            switch (command)
+            {
+                case nameof(Commands.SubsidiaryAdd):
+                    SubsidiaryAdd(stream, message, clientIp);
+                    break;
+                case nameof(Commands.SubsidiaryLoad):
+                    SubsidiaryLoad(stream);
+                    break;
+                default:
+                    break;
+            }
+        }
+
+        private static void SubsidiaryLoad(NetworkStream stream)
+        {
+            string Subsidiarys = "";
+            if (File.Exists(SubsidiaryManager.FilePath))
+                foreach (var item in File.ReadAllLines(SubsidiaryManager.FilePath))
+                    Subsidiarys += item + NetManager.separator;
+
+            NetManager.Send(stream, Subsidiarys);
+        }
+
+        private static void SubsidiaryAdd(NetworkStream stream, string message, string clientIp)
+        {
+            SubsidiaryManager.Add(message);
+            NetManager.Send(stream, String.Format("Добавлен новый филиал: {0}.", message));
+            Console.WriteLine("{0}: Добавлен новый филиал: {1}.", clientIp, message);
+        }
+    }
+}
